requirements.txt
#openai
#tiktoken
#langchain
#bardapi
#cmake
#python-poppler
#unstructured[pdf]
postgres
psycopg2
#mariadb-connector-c
mariadb
PyMySQL
pandas
numpy
sqlalchemy
vecs
pgvector
#gunicorn
#Flask
pyyaml

